+++ lib/rbush.dart
@@ -1,3 +1,3 @@
-export 'src/rbush.dart' show RBush, RBushBase, RBushBox, RBushElement;
+export 'src/rbush.dart' show RBush, RBushBase, RBushBox, RBushElement, RBushDirect;
 export 'src/quickselect.dart' show quickSelect;
 export 'src/tinyqueue.dart' show TinyQueue;