+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* `RBushElement` is now templated and requires a non-null data.
+* Added `RBushDirect<T>` for skipping bounding boxes management.
+
 ## 1.0.0
 
-- Initial version.
+* Initial version.